apps/functions-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "functions-python"
version = "0.1.0"
authors = [
    {name = "suzumina.click Team", email = "your.email@example.com"},
]
description = "Python Cloud Functions for suzumina.click - YouTube API integration"
requires-python = ">=3.12"
dependencies = [
    "functions-framework==3.5.0",
    "google-api-python-client==2.124.0",
    "google-auth==2.34.0",
    "google-cloud-secret-manager==2.22.0",
    "google-cloud-firestore==2.15.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.4.3",
    "pytest==8.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"